Push in ert_workflow_list into res_config
