+++ src/ert/cli/workflow.py
@@ -3,9 +3,8 @@ import logging
 
 def execute_workflow(ert, workflow_name):
     logger = logging.getLogger(__name__)
-    workflow_list = ert.getWorkflowList()
     try:
-        workflow = workflow_list[workflow_name]
+        workflow = ert.resConfig().workflows[workflow_name]
     except KeyError:
         msg = "Workflow {} is not in the list of available workflows"
         logger.error(msg.format(workflow_name))
