+++ src/ert/gui/tools/workflows/workflows_tool.py
@@ -8,7 +8,7 @@ class WorkflowsTool(Tool):
     def __init__(self, ert, notifier):
         self.notifier = notifier
         self.ert = ert
-        enabled = len(ert.getWorkflowList().getWorkflowNames()) > 0
+        enabled = len(ert.resConfig().workflow_jobs) > 0
         super().__init__(
             "Run workflow",
             "tools/workflows",
